As a player I want a visual indicator so that I can see my hunger level and manage it.
Merged thirstbar script into resource storage script and implemented visual indicator into tech demo

+++ Assets/Scripts/Units/UnitController.cs
@@ -13,7 +13,7 @@ public class UnitController : MonoBehaviour {
     public GameObject unitBase;
 
     public int woodCollected;
-    public int hungerCollected;
+    public float hungerCollected;
 
 	void Awake () {
         agent = GetComponent<NavMeshAgent>();
